Add bash script runner
This also removes the hard-coded venv path from the cmod
script itself, which is a nice decoupling. This also frees
up the user to use a system-level installation if desired,
or even a venv somewhere else on the system.

Still need to make a Windows .bat script to do this same thing