Redo arg parser
Raw args are now passed directly to
each command handler to allow runtime
argument generation for the cmod generator.

This signals a need to cleanup the CmodCommand
class among things down in that neighborhood.

Generator class still needs to be defined and moved
to it's owm module.

The generate handler can now handle custom generate flags
as arguments as read from the config file. The handler
can also currently distinguish between generation of
a list of individual files, and generating a pre-defined
module a described in the config.

We are getting close...